+++ src/components/roadmap/manage/member/BanUserModal/style.ts
@@ -1,4 +1,3 @@
-import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 
 export const Title = styled.h3`
