+++ src/components/roadmap/manage/member/Table/index.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useRouter } from 'next/router';
 import { useDeleteRoadmapGroupMember, useGetRoadmapGroupMember } from '@/api/hooks/roadmap';
-import BanUserModal from '@/components/roadmap/manage/BanUserModal';
+import BanUserModal from '@/components/roadmap/manage/member/BanUserModal';
 import TableColumn from '@/components/roadmap/manage/member/TableColumn';
 import { useModalState } from '@/hooks/useModalState';
 import * as Styled from './style';
